+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## v0.1.0 (2024-06-17)
+## v0.1.2 (2024-07-23)
+* Fixed int coercion in tryFromAny()
+* Fixed string comparison with ucfirst(name)
+
+## v0.1.1 (2024-06-17)
 * Added fromAny and tryFromAny
 
 ## v0.1.0 (2024-06-14)